docs(example): quarantine foreign plan workaround
Move the temporary execution-plan token registry into a deliberately named module and document its process-local behavior and removal criteria.
